# api-zod__generated_types_categoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}