# api-zod__generated_types_customersPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from './customer';

export interface CustomersPage {
  customers: Customer[];
  total: number;
  page: number;
  limit: number;
}